Fix all issues reported by check()
 * Incorrect formatting of license field in DESCRIPTION
 * fix parameter bug in vlift
 * remove magrittr dependency

check() reports a NOTE when `%>%` is used from magrittr even if magrittr
is imported in DESCRIPTION. The reason is that it expects every function
call to have an explicit namespace (e.g. dplyr::filter). A NOTE is not a
serious issue and will not cause Travis to fail, but for now I will just
content myself with no magrittr. In fact, since I already have my own
composition handling implemented, I don't even need the library.