Give inode argument to hnode a default value
Previously inode was a required argument. However, a user should be able
to manipulate a node before loading it into a graph context.

Due to lexical typing and non-reference sematics in R, any modification
to input functions, such as resetting a validator with h_val, will
create a new object and not be transferred to pre-existing inode lists.
So it is best to load last.

This is a serious problem. Currently I cannot map across my function
graph and swap out components without breaking links. Furthermore, since
hnodes do not hold links to downstream nodes, I cannot invalidate nodes
I break when I alter a function.