Use substitute in argument assignments in hnode
Before
```
R> hnode(foo)
function(.fun = f, ...
  ...
```

f is evaluated in the environment of _hsource or _hpipe, and is a
constant equal to the input function f.

Now
```
R> hnode(foo)
function(.fun = foo, ...
  ...
```

R will now search up the environment tree to find `foo`, which it will
find in the global environment. This search will occur every time the
function is called, possibly resulting in different values if `foo` is
redefined -- dynamic scoping. This can be dangerous, but is powerful.

Perhaps I should add a `protect` function that freezes an hnode by
sweeping all formal arguments into the body.