Allow runall to return a value
Originally I had it run all input functions, purely for side effects,
and return nothing. But this isn't really necessary. When runall is used
to process the effect list of functions in a node, it's returned values
are already ignored. There may be cases where it would be convenient for
the output to be automatically printed, so I might as well allow for
this.