Add evals to hnode
All the `substitute` wrappings allowed highly informative printing of
functions (preserved the user provided function names) and dynamic
scoping. However, they also did not run. The reason was that the input
functions were not functions but *name*s. I had thought these names were
still callable (autocasting magic or something), but I was wrong.

So I have to run eval on every substituted argument. This is terribly
ugly, but it fixes the problem.