Add warn and fail, remove warn and error
Previously I defined warn and error as wrappers of warning and stop that
passed a string and ... into sprintf before going into the warning/stop
message. It saved a few characters of typing, but also screwed with the
output and reduced readability by veering from expectations. So I
removed both functions.

I also added a warn and fail functions to atomic.R. Here the simply
trigger a warning or exception for any input. These will be useful as
default functions or in my tests.