Preserve srcref across extractions
Extract operations, such as occur in all my `h_*<-` functions, do not
preserve attributes, so I have been wrapping these functions with

```
a <- attributes(h)
formals(h)$x <- 'whatever'
attributes(h) <- a
```

It is ugly, but it works. However, I found that discrepencies between my
defaults shown when I print a functions, and the actual defaults
obtained from `formal(h)`. This was one of those disturbing bugs where
everything works fine but shouldn't. All my tests passed, but it seemed,
based off the defaults shown when I printed the functions, that they
shouldn't. Ultimately, I found the problem was that the function string
that is printed *is* an attribute, or at least sometimes it is. For some
functions, there is a 'srcref' attribute which is printed in lieu of the
deparsed function expression. This attribute is automatically generated,
under some conditions, it comes and goes ... or so it seems.

There are two solutions to this problem:

remove the srcref attribute

```
a <- attributes(h)
formals(h)$x <- 'whatever'
attributes(h) <- a
attributes(h)$srcref <- NULL
```

or add the new srcref

```
a <- attributes(h)
formals(h)$x <- 'whatever'
srcref <- attributes(h)$srcref
attributes(h) <- a
attributes(h)$srcref <- srcref
```

I will go with the former for now.