Release v0.1.0
From here on I will separate dev and master; every commit to master will
be a new release with a new version number.

This release is a weakly stable implementation of the basic needed
functionality. Among major omissions are control structures (loops and
if-statements), type searching, spiders, layers, and true manifolds.