Make hwell/pipe/sink return node builders
Previously the hnode functions (hwell, hpipe, and hsink) directly
returned a node of a given type. This node could be edited, giving it a
pure function, a validator, input nodes, etc. Then it could be called.
But there is a serious problem with this approach.

If a given operation occur multiple times in a pipeline, or is called on
multiple inputs (e.g. in a loop), each one must be uniquely identifiable
and have a unique cache. However, the hnode functions built a single
node. There is no clean way to copy these functions or reuse them. This
totally foobars looping, sharing and reuse.

My solution is to have the hnode function instead return a node maker.
This node maker can be customized (exactly as the nodes were
previously), then called to create new nodes with unique caches and
identities, but otherwise identical. These node makers are suitable for
export as libraries of modular, type-searchable, pipeline components.

Input nodes cannot be specified for a node maker. They are built in
isolation. When `connect` is called, node makers are evaluated to create
the required nodes, the nodes are connected, and then a pipeline object
is returned.

This commit implements half of the above. The build is broken. Here is
what remains to be done:

 * rewrite `connect`
 * sync tests
 * sync vignette
 * solve unanticipated problems ...